+++ apps/rust/test_app/src/primes.rs
@@ -6,11 +6,11 @@
 // Inline attributes are just to be on pair with C++ implementation (see primes.hpp)
 
 #[inline]
-fn is_prime(n: i64) -> bool {
+fn is_prime(n: i32) -> bool {
     if n <= 1 {
         return false;
     }
-    
+
     if n == 2 {
         return true;
     }
@@ -18,10 +18,10 @@ fn is_prime(n: i64) -> bool {
     // the only even prime is 2
     if n % 2 == 0 {
         return false;
-    } 
+    }
 
     // Trial division (https://en.wikipedia.org/wiki/Prime_number#Trial_division)
-    let limit = (n as f64).sqrt() as i64 + 1;
+    let limit = (n as f32).sqrt() as i32 + 1;
 
     for i in 3..limit {
         if n % i == 0 {
@@ -33,8 +33,8 @@ fn is_prime(n: i64) -> bool {
 }
 
 #[inline]
-pub fn get_primes_less_than(limit: i64) -> Vec<i64> {
-    let mut result: Vec<i64> = Vec::new();
+pub fn get_primes_less_than(limit: i32) -> Vec<i32> {
+    let mut result: Vec<i32> = Vec::new();
 
     for i in 2..limit {
         if is_prime(i) {
@@ -46,8 +46,8 @@ pub fn get_primes_less_than(limit: i64) -> Vec<i64> {
 }
 
 #[inline]
-pub fn count_primes_less_than(limit: i64) -> i64 {
-    let mut result: i64 = 0;
+pub fn count_primes_less_than(limit: i32) -> i32 {
+    let mut result: i32 = 0;
 
     for i in 2..limit {
         if is_prime(i) {
